internal/natsort: rename sortorder to natsort and remove vanity import
