--- internal/natsort/doc.go
@@ -1,5 +1,0 @@
-// Package sortorder implements sort orders and comparison functions.
-//
-// Currently, it only implements so-called "natural order", where integers
-// embedded in strings are compared by value.
-package sortorder // import "vbom.ml/util/sortorder"
